Check file is there before deleting it

+++ inc/fitbit.php
@@ -522,8 +522,10 @@
                     if (count($cacheNames) > 0) {
                         foreach ($cacheNames as $cacheName) {
                             if (substr($file, 0, strlen($username) + strlen($cacheName) + 2) === "_" . $username . "_" . $cacheName) {
-                                nxr("  $file cache file was deleted");
-                                unlink($cache_dir . $file);
+                                if (file_exists($cache_dir . $file) && is_writable($cache_dir . $file)) {
+                                    nxr("  $file cache file was deleted");
+                                    unlink($cache_dir . $file);
+                                }
                             }
                         }
                     }
